Add reminder for Deadline and Event classes
Streams are used in the tasklist to filter out tasks that are not upcoming. A isUpcoming method method returning a boolean is used to check if the deadline or event object is upcoming. RemindCommand returns a string listing all the upcoming tasks.

This adheres to "Ask, don't tell" principle as the TaskList outputs therelevant content requested by the user (and parsed). A RemindCommand is created separately, extending the command class, so that when Leo.java is run, polymorphsim allows c.execute(tasks, ui, storage) to exhibit different behaviour`